Circular Linked List Implementation
Create a circular linked list . 
isEmpty() : Checks if the list cLL is empty return boolean : true if empty ,false if not.
push(int xData) : Adds data to the beginning of the list
append(int xData) : Adds data to the end of the list 
pollHead() : Deletes the head (first node ) 
pollTail() : Deletes the tail (lastnode ) 
contains(int xData) : Returns boolean type. TRUE if the data is in the list ,FALSE if not.
length() ; Return the length of the list (number of nodes) .
delete(int xData): Deletes a node with the inputed xData. 
printList(): Prints the list.